# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  handlers.AppleOAuthRequest:
    properties:
      device_info:
        type: string
      id_token:
        type: string
    required:
    - id_token
    type: object
  handlers.FacebookOAuthRequest:
    properties:
      access_token:
        type: string
      device_info:
        type: string
    required:
    - access_token
    type: object
  handlers.GoogleOAuthRequest:
    properties:
      device_info:
        type: string
      id_token:
        type: string
    required:
    - id_token
    type: object
...
  models.AuthResponse:
    properties:
      mfa_challenge_id:
        type: string
      profile:
        $ref: '#/definitions/models.ProfileResponse'
      requires_mfa:
        type: boolean
      tokens:
        $ref: '#/definitions/models.TokenPair'
      user:
        $ref: '#/definitions/models.UserResponse'
    type: object
...
  models.CreatePollRequest:
    properties:
      options:
        items:
          type: string
        maxItems: 10
        minItems: 2
        type: array
    required:
    - options
    type: object
...
  models.Photo:
    properties:
      height:
        type: integer
      mime_type:
        type: string
      name:
        type: string
      size:
        type: integer
      url:
        type: string
      width:
        type: integer
    type: object
  models.PollOptionResponse:
    properties:
      id:
        type: string
      option:
        type: string
      percentage:
        type: number
      vote_count:
        type: integer
    type: object
  models.PollResponse:
    properties:
      created_at:
        type: string
      has_voted:
        type: boolean
      id:
        type: string
      options:
        items:
          $ref: '#/definitions/models.PollOptionResponse'
        type: array
      post_id:
        type: string
      total_votes:
        type: integer
      updated_at:
        type: string
      user_vote:
        description: Poll option ID that user voted for
        type: string
    type: object
...
  models.ProfileResponse:
    properties:
      avatar:
        $ref: '#/definitions/models.Photo'
      first_name:
        type: string
      id:
        type: string
      is_complete:
        type: boolean
      last_name:
        type: string
    type: object
...
  models.TokenPair:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      mfa_enabled:
        type: boolean
      phone_verified:
        type: boolean
    type: object
...
  models.VotePollRequest:
    properties:
      poll_option_id:
        type: string
    required:
    - poll_option_id
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hamsaya.app
    name: API Support
    url: https://hamsaya.app/support
  description: A production-ready Go backend for a social media mobile application
    with posts, business profiles, marketplace, real-time chat, and location services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://hamsaya.app/terms/
  title: Hamsaya Backend API
  version: "1.0"
paths:
...
  /auth/oauth/apple:
    post:
      consumes:
      - application/json
      description: Authenticate or register a user using Apple OAuth
      parameters:
      - description: Apple OAuth request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AppleOAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Apple OAuth authentication
      tags:
      - auth
  /auth/oauth/facebook:
    post:
      consumes:
      - application/json
      description: Authenticate or register a user using Facebook OAuth
      parameters:
      - description: Facebook OAuth request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.FacebookOAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Facebook OAuth authentication
      tags:
      - auth
  /auth/oauth/google:
    post:
      consumes:
      - application/json
      description: Authenticate or register a user using Google OAuth
      parameters:
      - description: Google OAuth request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.GoogleOAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Google OAuth authentication
      tags:
      - auth
...
  /polls/{poll_id}:
    get:
      description: Get a poll by ID
      parameters:
      - description: Poll ID
        in: path
        name: poll_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PollResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get a poll
      tags:
      - polls
  /polls/{poll_id}/vote:
    delete:
      description: Remove user's vote from a poll
      parameters:
      - description: Poll ID
        in: path
        name: poll_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete vote
      tags:
      - polls
    post:
      consumes:
      - application/json
      description: Vote on a poll option
      parameters:
      - description: Poll ID
        in: path
        name: poll_id
        required: true
        type: string
      - description: Vote request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.VotePollRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PollResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Vote on a poll
      tags:
      - polls
...
  /posts/{post_id}/polls:
    get:
      description: Get poll by post ID
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PollResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get poll for a post
      tags:
      - polls
    post:
      consumes:
      - application/json
      description: Create a poll for a PULL type post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      - description: Poll creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreatePollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.PollResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create a poll
      tags:
      - polls
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"